requirements.txt
absl-py==1.3.0
aiohttp==3.8.3
aiosignal==1.3.1
antlr4-python3-runtime==4.8
async-timeout==4.0.2
attrs==22.1.0
beautifulsoup4==4.11.1
bitarray==2.6.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
contourpy==1.0.6
cycler==0.11.0
Cython==0.29.32
datasets==2.7.1
dill==0.3.6
einops==0.6.0
-e ./OFA/OFA/fairseq
filelock==3.8.0
fire==0.4.0
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2022.11.0
ftfy==6.0.3
gdown==4.7.1
html5lib==1.1
huggingface-hub==0.11.1
hydra-core==1.0.7
idna==3.4
imageio==2.22.4
joblib==1.2.0
kiwisolver==1.4.4
Levenshtein==0.20.8
lightning-utilities==0.3.0
lxml==4.9.1
matplotlib==3.6.2
metrics==0.3.3
multidict==6.0.2
multiprocess==0.70.14
networkx==2.8.8
nltk==3.7
numpy==1.23.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
omegaconf==2.0.6
opencv-python==4.6.0.66
packaging==21.3
pandas==1.5.1
pathlib2==2.3.7.post1
pathspec==0.5.5
Pillow==9.3.0
portalocker==2.6.0
protobuf==3.20.0
pyarrow==10.0.1
pycocoevalcap==1.2
pycocotools==2.0.4
pycparser==2.21
Pygments==2.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-Levenshtein==0.20.8
pytorch-lightning==1.8.3.post1
pytz==2022.5
PyWavelets==1.4.1
PyYAML==6.0
rapidfuzz==2.13.2
regex==2022.10.31
requests==2.28.1
responses==0.18.0
rouge-score==0.1.2
sacrebleu==2.3.1
scikit-image==0.19.3
scikit-learn==1.2.1
scipy==1.9.3
six==1.16.0
sklearn==0.0.post1
soupsieve==2.3.2.post1
tabulate==0.9.0
tensorboardX==2.4.1
termcolor==2.1.1
threadpoolctl==3.1.0
tifffile==2022.10.10
timm==0.6.12
torch==1.13.0
torchmetrics==0.11.0
torchvision==0.14.0
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.12
wcwidth==0.2.5
webencodings==0.5.1
xxhash==3.1.0
yarl==1.8.1
zhconv==1.4.3
